pyproject.toml
[project]
name = "django-backend-template"
version = "0.1.0"
description = "A clean and extensible Django project starter template with batteries included. Built to help you bootstrap new Django applications quickly with best practices."
authors = [
    {name = "Murad",email = "nuradhussen082@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django (>=5.2.4,<6.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "drf-spectacular[sidecar] (>=0.28.0,<0.29.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)",
    "daphne (>=4.2.1,<5.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "django-filter (>=25.1,<26.0)",
    "djangorestframework-simplejwt[crypto] (>=5.5.1,<6.0.0)",
    "pre-commit (>=4.2.0,<5.0.0)",
    "celery (>=5.5.3,<6.0.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "channels (>=4.3.1,<5.0.0)",
    "pillow (>=12.0.0,<13.0.0)",
    "whitenoise (>=6.11.0,<7.0.0)"
]

[tool.poetry]
package-mode = false
[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.1"
flake8 = "^7.3.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
